Add support for whatwg streams
Add static methods to TextEncoder and TextDecoder allowing them to vend
TransformStream instances. These can then interface with other APIs
based on the whatwg streams specification. A stream of strings can be
converted to a byte stream, or a byte stream can be converted to a
stream of strings.

The streams spec can be found at https://streams.spec.whatwg.org/

TransformStream is not yet specified, but a reference implementation
exists. This change is not up-to-date with the latest version of
the reference implementation API. The API is not yet stable.