Use anonymous transformer objects
Previously the stream() methods were implemented in terms of
TextEncoderTransformer and TextDecoderTransformer objects. However,
these interfaces weren't intended to be exposed to web developers. They
added a lot of unnecessary words to the spec.

This change makes the transformer objects be unnamed. Methods are
defined on the object directly.

This change is invisible to users of the API.